+++ src/main/java/pl/plajer/murdermystery/api/events/player/package-info.java
@@ -13,29 +13,8 @@
  * along with Murder Mystery.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package pl.plajer.murdermystery.murdermysteryapi;
-
-import org.bukkit.event.Event;
-
-import pl.plajer.murdermystery.arena.Arena;
 
 /**
- * Represents Murder Mystery game related events.
+ * All in-game player related events.
  */
-public abstract class MurderMysteryEvent extends Event {
-
-  protected Arena arena;
-
-  public MurderMysteryEvent(Arena eventArena) {
-    arena = eventArena;
-  }
-
-  /**
-   * Returns event arena
-   *
-   * @return event arena
-   */
-  public Arena getArena() {
-    return arena;
-  }
-}
+package pl.plajer.murdermystery.api.events.player;