+++ src/main/java/pl/plajer/murdermystery/events/JoinEvent.java
@@ -16,7 +16,6 @@
 package pl.plajer.murdermystery.events;
 
 import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -25,10 +24,9 @@ import org.bukkit.event.player.PlayerLoginEvent;
 
 import pl.plajer.murdermystery.Main;
 import pl.plajer.murdermystery.arena.ArenaRegistry;
-import pl.plajer.murdermystery.database.FileStats;
 import pl.plajer.murdermystery.database.MySQLConnectionUtils;
 import pl.plajer.murdermystery.handlers.PermissionsManager;
-import pl.plajer.murdermystery.murdermysteryapi.StatsStorage;
+import pl.plajer.murdermystery.api.StatsStorage;
 import pl.plajer.murdermystery.user.UserManager;
 import pl.plajerlair.core.services.exception.ReportedException;