+++ src/main/java/pl/plajer/murdermystery/handlers/PlaceholderManager.java
@@ -19,7 +19,7 @@ import me.clip.placeholderapi.expansion.PlaceholderExpansion;
 
 import org.bukkit.entity.Player;
 
-import pl.plajer.murdermystery.murdermysteryapi.StatsStorage;
+import pl.plajer.murdermystery.api.StatsStorage;
 
 /**
  * @author Plajer