+++ src/main/java/pl/plajer/murdermystery/user/User.java
@@ -27,9 +27,8 @@ import org.bukkit.scoreboard.ScoreboardManager;
 import pl.plajer.murdermystery.Main;
 import pl.plajer.murdermystery.arena.Arena;
 import pl.plajer.murdermystery.arena.ArenaRegistry;
-import pl.plajer.murdermystery.database.FileStats;
-import pl.plajer.murdermystery.murdermysteryapi.MMPlayerStatisticChangeEvent;
-import pl.plajer.murdermystery.murdermysteryapi.StatsStorage;
+import pl.plajer.murdermystery.api.events.player.MMPlayerStatisticChangeEvent;
+import pl.plajer.murdermystery.api.StatsStorage;
 
 /**
  * @author Plajer