+++ src/main/java/pl/plajer/murdermystery/arena/ArenaManager.java
@@ -242,7 +242,7 @@ public class ArenaManager {
 
       p.setGameMode(GameMode.SURVIVAL);
       for (Player players : plugin.getServer().getOnlinePlayers()) {
-        if (ArenaRegistry.getArena(players) != null) {
+        if (ArenaRegistry.getArena(players) == null) {
           players.showPlayer(p);
         }
         p.showPlayer(players);