+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added spectator settings
 * API recode
 * JavaDocs created
+* Fixed throwable sword angle
 
 ### 0.0.7 Beta (03.10.2018)
 * Added PAPI placeholders in chat (chat formatting must be enabled)