+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fixed throwable sword angle
 * You cannot join game now infinitely through command
 * Added corpses override option in config
+* Now players that will leave game will be visible by other players outside game
 
 ### 0.0.7 Beta (03.10.2018)
 * Added PAPI placeholders in chat (chat formatting must be enabled)